packages/agentic_event_orchestrator/requirements.txt
pydantic
python-dotenv
requests
chainlit
ortools
google-generativeai
psycopg2-binary
mcp
